refactoring the event code into different directory
