+++ src/events/handlers/damageHandler.ts
@@ -1,35 +1,35 @@
 import type { TAudioManager } from '@/createAudioManager';
 import { flashRed } from '@/createEffectManager';
-import type { ECSEmitter } from '@/createGameLoop';
 import type { ECSEntityId } from '@/ecs/ECSEntity';
 import type { ECSWorld } from '@/ecs/ECSWorld';
 import { deleteComponent } from '@/entity/components/Delete';
 import { hasEnemy } from '@/entity/components/Enemy';
 import { hasPlayer } from '@/entity/components/Player';
 import { dealDamage } from '@/utils/ecsUtils';
+import type { ECSEmitter } from '../createExternalQueue';
 
 export const damageHandler = (
   options: {
     damage: number;
     entityId: ECSEntityId;
   },
-  ecs: ECSWorld,
+  world: ECSWorld,
   navigateTo: (path: string) => void,
   emit: ECSEmitter
 ) => {
-  const entityToDamage = ecs.getEntity(options.entityId).unwrap();
+  const entityToDamage = world.getEntity(options.entityId).unwrap();
 
   let stats: any;
 
   dealDamage({
     to: entityToDamage as any,
     amount: options.damage,
-    ecs
+    world: world
   });
 
   if (hasPlayer(entityToDamage)) {
     ({ stats } = entityToDamage.player);
-    ecs.get<TAudioManager>('audio').unwrap().play('ouch');
+    world.get<TAudioManager>('audio').unwrap().play('ouch');
     emit('playerHealthChanged');
     if (stats.current.health <= 0) {
       alert('you dead');
@@ -37,11 +37,11 @@ export const damageHandler = (
     }
   } else if (hasEnemy(entityToDamage)) {
     ({ stats } = entityToDamage.enemy);
-    ecs.get<TAudioManager>('audio').unwrap().play('damage');
+    world.get<TAudioManager>('audio').unwrap().play('damage');
   }
 
   if (stats.current.health <= 0) {
-    ecs.addComponent(entityToDamage, deleteComponent);
+    world.addComponent(entityToDamage, deleteComponent);
   }
 
   flashRed(entityToDamage.entity_id);
