+++ src/events/handlers/highlightInteractablesHandler.ts
@@ -2,7 +2,7 @@ import type { ECSWorld } from '@/ecs/ECSWorld';
 
 export const highlightInteractablesHandler = (
   isHighlightEnabled: boolean,
-  ecs: ECSWorld
+  world: ECSWorld
 ) => {
-  ecs.set('highlightInteractables', isHighlightEnabled);
+  world.set('highlightInteractables', isHighlightEnabled);
 };
